al-study-buddy/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
